ai-service/requirements.txt
flask==2.3.3
scikit-learn==1.3.0
tensorflow==2.13.0
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
joblib==1.3.1
